[UPDATE/FIXBUG/MERGE]完成客观题历史业务层、完善考试内容、合并上游分支

+++ dcoj-web/src/main/java/com/dcoj/controller/ObjectJudgeController.java
@@ -37,7 +37,7 @@ public class ObjectJudgeController {
         int uid = JWTUtil.getUid(token);
         int i = objectProblemService.judgeObjectProblem(userObjectJudgeFormat.getPid(), userObjectJudgeFormat.getAnswer());
         // TODO Leon 20190528：更新客观题状态
-        objectProblemUserService.insertOrUpdate(userObjectJudgeFormat.getPid(), uid, i);
+        objectProblemUserService.insertOrUpdate(userObjectJudgeFormat.getPid(), uid, (byte) i);
         // TODO Leon 20190528：Update User Log
         return null;
     }
